Finalize native EVMYulLean report boundary
