+++ .github/workflows/evmyullean-fork-conformance.yml
@@ -22,11 +22,11 @@ on:
       - '.github/workflows/evmyullean-fork-conformance.yml'
       - 'Makefile'
       - 'Compiler/Proofs/EndToEnd.lean'
-      - 'scripts/generate_evmyullean_adapter_report.py'
+      - 'scripts/generate_evmyullean_native_lowering_report.py'
       - 'scripts/generate_evmyullean_fork_audit.py'
       - 'scripts/test_evmyullean_fork_conformance_workflow.py'
       - 'artifacts/evmyullean_fork_audit.json'
-      - 'artifacts/evmyullean_adapter_report.json'
+      - 'artifacts/evmyullean_native_lowering_report.json'
       - 'lake-manifest.json'
       - 'Compiler/Proofs/YulGeneration/Backends/EvmYulLeanNativeLowering.lean'
       - 'Compiler/Proofs/YulGeneration/Backends/EvmYulLeanBodyClosure.lean'
@@ -41,11 +41,11 @@ on:
       - '.github/workflows/evmyullean-fork-conformance.yml'
       - 'Makefile'
       - 'Compiler/Proofs/EndToEnd.lean'
-      - 'scripts/generate_evmyullean_adapter_report.py'
+      - 'scripts/generate_evmyullean_native_lowering_report.py'
       - 'scripts/generate_evmyullean_fork_audit.py'
       - 'scripts/test_evmyullean_fork_conformance_workflow.py'
       - 'artifacts/evmyullean_fork_audit.json'
-      - 'artifacts/evmyullean_adapter_report.json'
+      - 'artifacts/evmyullean_native_lowering_report.json'
       - 'lake-manifest.json'
       - 'Compiler/Proofs/YulGeneration/Backends/EvmYulLeanNativeLowering.lean'
       - 'Compiler/Proofs/YulGeneration/Backends/EvmYulLeanBodyClosure.lean'
