+++ .github/workflows/verify.yml
@@ -254,7 +254,7 @@ jobs:
           python3 scripts/generate_layer2_boundary_catalog.py
           python3 scripts/generate_verify_sync_spec.py
           python3 scripts/generate_evmyullean_capability_report.py
-          python3 scripts/generate_evmyullean_adapter_report.py
+          python3 scripts/generate_evmyullean_native_lowering_report.py
           python3 scripts/generate_print_axioms.py
           python3 scripts/sync_verification_status_doc.py
