+++ artifacts/COMPILER_ARTIFACTS.md
@@ -11,7 +11,7 @@ This folder collects compiler artifacts and fixtures used by tests, CI, and docu
 - `lean_warning_baseline.json`: Lean warning regression baseline. Consumed by `scripts/check_lean_warning_regression.py`.
 - `interpreter_feature_matrix.json`: Machine-readable interpreter feature support matrix.
 - `evmyullean_capability_report.json`: EVMYulLean builtin coverage report. Generated by `scripts/generate_evmyullean_capability_report.py`.
-- `evmyullean_adapter_report.json`: EVMYulLean native lowering and bridge report.
+- `evmyullean_native_lowering_report.json`: EVMYulLean native lowering and bridge report.
 - `evmyullean_unsupported_nodes.json`: EVMYulLean unsupported AST node report.
 - `issue_1060_progress.json`: Issue #1060 migration tracking ledger (complete).
