+++ artifacts/evmyullean_native_lowering_report.json
@@ -182,7 +182,7 @@
     "runtime_lowering": "present"
   },
   "native_lowering_file": "Compiler/Proofs/YulGeneration/Backends/EvmYulLeanNativeLowering.lean",
-  "schema_version": 7,
+  "schema_version": 8,
   "status": "partial",
   "stmt_gap_messages": {
     "funcDef": [
