+++ scripts/refresh_verification_artifacts.sh
@@ -9,7 +9,7 @@ python3 scripts/generate_verification_status.py
 python3 scripts/generate_layer2_boundary_catalog.py
 python3 scripts/generate_verify_sync_spec.py
 python3 scripts/generate_evmyullean_capability_report.py
-python3 scripts/generate_evmyullean_adapter_report.py
+python3 scripts/generate_evmyullean_native_lowering_report.py
 python3 scripts/generate_print_axioms.py
 python3 scripts/sync_verification_status_doc.py
 
@@ -18,7 +18,7 @@ python3 scripts/generate_verification_status.py --check
 python3 scripts/generate_layer2_boundary_catalog.py --check
 python3 scripts/generate_verify_sync_spec.py --check
 python3 scripts/generate_evmyullean_capability_report.py --check
-python3 scripts/generate_evmyullean_adapter_report.py --check
+python3 scripts/generate_evmyullean_native_lowering_report.py --check
 python3 scripts/generate_print_axioms.py --check
 python3 scripts/check_verification_status_doc.py
 python3 scripts/check_layer2_boundary_catalog_sync.py
