+++ scripts/test_generate_evmyullean_native_lowering_report.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""Tests for generate_evmyullean_adapter_report.py.
+"""Tests for generate_evmyullean_native_lowering_report.py.
 
 Validates bridge-lemma sorry detection, builtin extraction, fail-fast
 on missing files, and end-to-end report consistency with the repo.
@@ -18,7 +18,7 @@ SCRIPT_DIR = Path(__file__).resolve().parent
 if str(SCRIPT_DIR) not in sys.path:
     sys.path.insert(0, str(SCRIPT_DIR))
 
-import generate_evmyullean_adapter_report as gen
+import generate_evmyullean_native_lowering_report as gen
 
 
 class ParseNativeLoweringCasesTests(unittest.TestCase):
@@ -580,8 +580,8 @@ class RepoArtifactConsistencyTests(unittest.TestCase):
         self.assertEqual(
             existing,
             rendered,
-            "Adapter report artifact is stale. Regenerate with:\n"
-            "  python3 scripts/generate_evmyullean_adapter_report.py",
+            "Native lowering report artifact is stale. Regenerate with:\n"
+            "  python3 scripts/generate_evmyullean_native_lowering_report.py",
         )
 
     def test_admitted_lemmas_are_subset_of_universal(self) -> None:
@@ -660,699 +660,13 @@ class RepoArtifactConsistencyTests(unittest.TestCase):
             f"Universal bridge lemmas not in bridged set: {universal - bridged}",
         )
 
-    @unittest.skipUnless(
-        gen.RETARGET_FILE.exists(),
-        "phase4 retarget tracking removed after EVMYulLean transition",
-    )
-    def test_scalar_parameter_body_closure_is_tracked(self) -> None:
+    def test_removed_transition_payload_is_absent(self) -> None:
         report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(
-            phase4["genParamLoads_scalar_bridged"],
-            "proven (scalar calldata parameters)",
-        )
-        self.assertEqual(
-            phase4["genStaticTypeLoads_calldataload_bridged"],
-            "proven (static scalar calldata leaves)",
-        )
-        self.assertEqual(
-            phase4["genParamLoads_static_scalar_bridged"],
-            "proven (static scalar calldata parameters)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_binding_leaf_bridged"],
-            "proven (scalar let/assign statement lists)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_pure_binding_bridged"],
-            "proven (pure let/assign statement lists)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_storage_fragment_bridged"],
-            "proven (pure bindings plus single-slot setStorage statement lists)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_terminator_external_bridged"],
-            "proven (external stop/return statement lists)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_internal_return_bridged"],
-            "proven (internal return statement lists)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_require_bridged"],
-            "proven (require statement lists)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_external_body_fragment_bridged"],
-            "proven (mixed external body fragment)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_internal_body_fragment_bridged"],
-            "proven (mixed internal body fragment)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_external_structured_body_fragment_bridged"],
-            "proven (mixed external body fragment plus one ite layer)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_internal_structured_body_fragment_bridged"],
-            "proven (mixed internal body fragment plus one ite layer)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_external_nested_body_fragment_bridged"],
-            "proven (mixed external body fragment plus two ite layers)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_internal_nested_body_fragment_bridged"],
-            "proven (mixed internal body fragment plus two ite layers)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_external_recursive_body_fragment_bridged"],
-            "proven (mixed external body fragment plus recursive ite closure)",
-        )
-        self.assertEqual(
-            phase4["compileStmtList_internal_recursive_body_fragment_bridged"],
-            "proven (mixed internal body fragment plus recursive ite closure)",
-        )
-        self.assertEqual(
-            phase4["compileExpr_bridgedSource_leaf"],
-            "proven (scalar source-expression leaves)",
-        )
-        self.assertEqual(
-            phase4["compileExpr_bridgedSource"],
-            "proven (source-expression fragment with parameter length, "
-            "storage, and storage-array length reads, ADT tag/field reads, "
-            "singleton and nested mapping reads through the abstract "
-            "mappingSlot bridge, mapping struct-member reads, "
-            "reserved exponentiation, boolean normalization, branchless "
-            "helpers, bridged environment reads, and unary calldata/memory/"
-            "transient reads)",
-        )
-        self.assertEqual(
-            phase4["compileExpr_mappingChain_bridgedSource"],
-            "proven (mappingChain source-expression wrapper through "
-            "the abstract mappingSlot bridge)",
-        )
+        self.assertNotIn("phase4_" + "retarget", report)
 
-    def test_missing_retarget_theorem_is_not_reported_proven(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    /-- Mentions backends_agree_on_bridged_builtins only in prose. -/
-                    theorem other_theorem : True := by trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "incomplete")
-        self.assertEqual(phase4["backends_agree_on_bridged_builtins"], "missing")
-        self.assertEqual(phase4["evalYulExpr_evmYulLean_eq_on_bridged"], "missing")
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_eq_on_bridged_straight_stmts"],
-            "missing",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_block_eq_on_bridged_straight_stmts"],
-            "missing",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_if_eq_on_bridged_body"],
-            "missing",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_switch_eq_on_bridged_cases"],
-            "missing",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_for_eq_on_bridged_parts"],
-            "missing",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_eq_on_bridged_target"],
-            "missing",
-        )
-        self.assertEqual(phase4["emitYul_runtimeCode_bridged"], "missing")
-        self.assertEqual(
-            phase4["emitYul_runtimeCode_evmYulLean_eq_on_bridged_bodies"],
-            "missing",
-        )
-        self.assertEqual(
-            phase4["yulCodegen_preserves_semantics_evmYulLeanBackend_via_reference_oracle"],
-            "missing",
-        )
-
-    def test_retarget_theorem_name_inside_string_is_not_reported_proven(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent('''\
-                    def decoy : String :=
-                      "theorem backends_agree_on_bridged_builtins : True := by sorry"
-
-                    theorem other_theorem : True := by trivial
-                '''),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "incomplete")
-        self.assertEqual(phase4["backends_agree_on_bridged_builtins"], "missing")
-
-    def test_sorry_retarget_theorem_downgrades_phase4_status(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    theorem backends_agree_on_bridged_builtins : True := by
-                      trivial
-
-                    theorem evalYulExpr_evmYulLean_eq_on_bridged : True := by
-                      sorry
-
-                    theorem execYulFuelWithBackend_eq_on_bridged_straight_stmts : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_block_eq_on_bridged_straight_stmts : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_if_eq_on_bridged_body : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_switch_eq_on_bridged_cases : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_for_eq_on_bridged_parts : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_eq_on_bridged_target : True := by
-                      trivial
-
-                    theorem emitYul_runtimeCode_bridged : True := by
-                      trivial
-
-                    theorem emitYul_runtimeCode_evmYulLean_eq_on_bridged_bodies : True := by
-                      trivial
-
-                    theorem yulCodegen_preserves_semantics_evmYulLeanBackend_via_reference_oracle : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                with patch.object(gen, "_parse_bridge_lemmas", return_value=(["add"], [])):
-                    report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "pointwise")
-        self.assertEqual(phase4["backends_agree_on_bridged_builtins"], "proven")
-        self.assertEqual(phase4["evalYulExpr_evmYulLean_eq_on_bridged"], "sorry")
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_eq_on_bridged_straight_stmts"],
-            "proven",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_block_eq_on_bridged_straight_stmts"],
-            "proven",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_if_eq_on_bridged_body"],
-            "proven",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_switch_eq_on_bridged_cases"],
-            "proven",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_for_eq_on_bridged_parts"],
-            "proven",
-        )
-        self.assertEqual(
-            phase4["execYulFuelWithBackend_eq_on_bridged_target"],
-            "proven",
-        )
-        self.assertEqual(
-            phase4["emitYul_runtimeCode_bridged"],
-            "proven (conditional on bridged IR bodies)",
-        )
-        self.assertEqual(
-            phase4["emitYul_runtimeCode_evmYulLean_eq_on_bridged_bodies"],
-            "proven (conditional on bridged IR bodies)",
-        )
-        self.assertEqual(
-            phase4["yulCodegen_preserves_semantics_evmYulLeanBackend_via_reference_oracle"],
-            "proven (conditional on bridged IR bodies)",
-        )
-        self.assertEqual(phase4["admitted_bridge_dependencies"], [])
-
-    def test_top_level_local_sorry_does_not_downgrade_prior_retarget_theorem(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    theorem backends_agree_on_bridged_builtins : True := by
-                      trivial
-
-                    local def helper : True := by
-                      sorry
-
-                    theorem evalYulExpr_evmYulLean_eq_on_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                with patch.object(gen, "_parse_bridge_lemmas", return_value=(["add"], [])):
-                    report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["backends_agree_on_bridged_builtins"], "proven")
-        self.assertEqual(phase4["evalYulExpr_evmYulLean_eq_on_bridged"], "proven")
-
-    def test_retarget_theorem_body_includes_where_helper_sorry(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    theorem backends_agree_on_bridged_builtins : True := by
-                      exact helper
-                    where
-                      helper : True := by
-                        sorry
-
-                    theorem evalYulExpr_evmYulLean_eq_on_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                with patch.object(gen, "_parse_bridge_lemmas", return_value=(["add"], [])):
-                    report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertIn("sorry", phase4["backends_agree_on_bridged_builtins"])
-        self.assertEqual(phase4["evalYulExpr_evmYulLean_eq_on_bridged"], "proven")
-
-    def test_opaque_between_retarget_theorems_is_own_boundary(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    theorem backends_agree_on_bridged_builtins : True := by
-                      trivial
-
-                    opaque helper : True := by
-                      sorry
-
-                    theorem evalYulExpr_evmYulLean_eq_on_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                with patch.object(gen, "_parse_bridge_lemmas", return_value=(["add"], [])):
-                    report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["backends_agree_on_bridged_builtins"], "proven")
-        self.assertEqual(phase4["evalYulExpr_evmYulLean_eq_on_bridged"], "proven")
-
-    def test_nested_local_theorem_does_not_satisfy_retarget_presence(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    theorem wrapper : True := by
-                      local theorem backends_agree_on_bridged_builtins : True := by
-                        trivial
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                with patch.object(gen, "_parse_bridge_lemmas", return_value=(["add"], [])):
-                    report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["backends_agree_on_bridged_builtins"], "missing")
-        self.assertEqual(phase4["status"], "incomplete")
-
-    def test_admitted_bridge_deps_downgrade_phase4_status(self) -> None:
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            retarget = Path(tmp) / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(
-                textwrap.dedent("""\
-                    theorem backends_agree_on_bridged_builtins : True := by
-                      trivial
-
-                    theorem evalYulExpr_evmYulLean_eq_on_bridged : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_eq_on_bridged_straight_stmts : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_block_eq_on_bridged_straight_stmts : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_if_eq_on_bridged_body : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_switch_eq_on_bridged_cases : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_for_eq_on_bridged_parts : True := by
-                      trivial
-
-                    theorem execYulFuelWithBackend_eq_on_bridged_target : True := by
-                      trivial
-
-                    theorem emitYul_runtimeCode_bridged : True := by
-                      trivial
-
-                    theorem emitYul_runtimeCode_evmYulLean_eq_on_bridged_bodies : True := by
-                      trivial
-
-                    theorem yulCodegen_preserves_semantics_evmYulLeanBackend_via_reference_oracle : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-            with patch.object(gen, "RETARGET_FILE", retarget):
-                with patch.object(
-                    gen,
-                    "_parse_bridge_lemmas",
-                    return_value=(["add", "smod", "sar"], ["smod", "sar"]),
-                ):
-                    report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "incomplete")
-        self.assertEqual(phase4["admitted_bridge_dependencies"], ["sar", "smod"])
-        self.assertIn("sorry-dependent", phase4["backends_agree_on_bridged_builtins"])
-        self.assertIn("smod", phase4["evalYulExpr_evmYulLean_eq_on_bridged"])
-        self.assertIn(
-            "smod",
-            phase4["execYulFuelWithBackend_eq_on_bridged_straight_stmts"],
-        )
-        self.assertIn(
-            "smod",
-            phase4["execYulFuelWithBackend_block_eq_on_bridged_straight_stmts"],
-        )
-        self.assertIn(
-            "smod",
-            phase4["execYulFuelWithBackend_if_eq_on_bridged_body"],
-        )
-        self.assertIn(
-            "smod",
-            phase4["execYulFuelWithBackend_switch_eq_on_bridged_cases"],
-        )
-        self.assertIn(
-            "smod",
-            phase4["execYulFuelWithBackend_for_eq_on_bridged_parts"],
-        )
-        self.assertIn(
-            "smod",
-            phase4["execYulFuelWithBackend_eq_on_bridged_target"],
-        )
-        self.assertEqual(
-            phase4["emitYul_runtimeCode_bridged"],
-            "proven (conditional on bridged IR bodies)",
-        )
-        self.assertIn(
-            "smod",
-            phase4["emitYul_runtimeCode_evmYulLean_eq_on_bridged_bodies"],
-        )
-        self.assertIn(
-            "smod",
-            phase4["yulCodegen_preserves_semantics_evmYulLeanBackend_via_reference_oracle"],
-        )
-
-    @unittest.skipUnless(
-        gen.RETARGET_FILE.exists(),
-        "phase4 retarget tracking removed after EVMYulLean transition",
-    )
-    def test_universal_body_closure_proven_in_repo(self) -> None:
-        """The repo artifact tracks the landed plan #1722 B7/B8 theorem stack."""
+    def test_schema_version_current(self) -> None:
         report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertIn("proven", phase4["compileStmtList_always_bridged"])
-        self.assertEqual(phase4["status"], "universal-safe-body-closure")
-        self.assertEqual(
-            phase4["layers2_3_ir_matches_yul_evmYulLeanBackend"],
-            "removed from EndToEnd surface",
-        )
-
-    def _retarget_all_proven(self) -> str:
-        return textwrap.dedent("""\
-            theorem backends_agree_on_bridged_builtins : True := by
-              trivial
-
-            theorem evalYulExpr_evmYulLean_eq_on_bridged : True := by
-              trivial
-
-            theorem execYulFuelWithBackend_eq_on_bridged_straight_stmts : True := by
-              trivial
-
-            theorem execYulFuelWithBackend_block_eq_on_bridged_straight_stmts : True := by
-              trivial
-
-            theorem execYulFuelWithBackend_if_eq_on_bridged_body : True := by
-              trivial
-
-            theorem execYulFuelWithBackend_switch_eq_on_bridged_cases : True := by
-              trivial
-
-            theorem execYulFuelWithBackend_for_eq_on_bridged_parts : True := by
-              trivial
-
-            theorem execYulFuelWithBackend_eq_on_bridged_target : True := by
-              trivial
-
-            theorem emitYul_runtimeCode_bridged : True := by
-              trivial
-
-            theorem emitYul_runtimeCode_evmYulLean_eq_on_bridged_bodies : True := by
-              trivial
-
-            theorem yulCodegen_preserves_semantics_evmYulLeanBackend_via_reference_oracle : True := by
-              trivial
-        """)
-
-    def test_universal_body_closure_flips_phase4_status(self) -> None:
-        """With all retarget theorems proven, no admitted bridge deps, an
-        unconditional public EndToEnd theorem, AND the B7 universal body
-        closure theorem proven, phase4_status must flip to
-        'full_semantic_integration'."""
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            tmp_path = Path(tmp)
-            retarget = tmp_path / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(self._retarget_all_proven(), encoding="utf-8")
-
-            end_to_end = tmp_path / "EndToEnd.lean"
-            end_to_end.write_text(
-                textwrap.dedent("""\
-                    theorem layers2_3_ir_matches_yul_evmYulLeanBackend : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            body_closure = tmp_path / "EvmYulLeanBodyClosure.lean"
-            body_closure.write_text(
-                textwrap.dedent("""\
-                    theorem compileStmtList_always_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            with patch.object(gen, "RETARGET_FILE", retarget), \
-                 patch.object(gen, "END_TO_END_FILE", end_to_end), \
-                 patch.object(gen, "BODY_CLOSURE_FILE", body_closure), \
-                 patch.object(
-                     gen,
-                     "_parse_bridge_lemmas",
-                     return_value=(["add"], []),
-                 ):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "full_semantic_integration")
-        self.assertIn(
-            "proven",
-            phase4["compileStmtList_always_bridged"],
-        )
-        self.assertEqual(phase4["admitted_bridge_dependencies"], [])
-
-    def test_universal_body_closure_with_sorry_does_not_flip_phase4_status(self) -> None:
-        """If compileStmtList_always_bridged is declared but has a sorry, the
-        phase4 status must stay at its current ceiling rather than advancing
-        to 'full_semantic_integration'."""
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            tmp_path = Path(tmp)
-            retarget = tmp_path / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(self._retarget_all_proven(), encoding="utf-8")
-
-            end_to_end = tmp_path / "EndToEnd.lean"
-            end_to_end.write_text(
-                textwrap.dedent("""\
-                    theorem layers2_3_ir_matches_yul_evmYulLeanBackend : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            body_closure = tmp_path / "EvmYulLeanBodyClosure.lean"
-            body_closure.write_text(
-                textwrap.dedent("""\
-                    theorem compileStmtList_always_bridged : True := by
-                      sorry
-                """),
-                encoding="utf-8",
-            )
-
-            with patch.object(gen, "RETARGET_FILE", retarget), \
-                 patch.object(gen, "END_TO_END_FILE", end_to_end), \
-                 patch.object(gen, "BODY_CLOSURE_FILE", body_closure), \
-                 patch.object(
-                     gen,
-                     "_parse_bridge_lemmas",
-                     return_value=(["add"], []),
-                 ):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertNotEqual(phase4["status"], "full_semantic_integration")
-        self.assertEqual(phase4["compileStmtList_always_bridged"], "sorry")
-
-    def test_body_closure_with_conditional_end_to_end_does_not_flip_to_full(self) -> None:
-        """B7 alone is not the final status: B8 must remove explicit
-        BridgedStmts body hypotheses from the public EndToEnd theorem."""
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            tmp_path = Path(tmp)
-            retarget = tmp_path / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(self._retarget_all_proven(), encoding="utf-8")
-
-            end_to_end = tmp_path / "EndToEnd.lean"
-            end_to_end.write_text(
-                textwrap.dedent("""\
-                    theorem layers2_3_ir_matches_yul_evmYulLeanBackend
-                        (hFunctions : BridgedStmts body)
-                        (hInternals : BridgedStmts internals) : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            body_closure = tmp_path / "EvmYulLeanBodyClosure.lean"
-            body_closure.write_text(
-                textwrap.dedent("""\
-                    theorem compileStmtList_always_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            with patch.object(gen, "RETARGET_FILE", retarget), \
-                 patch.object(gen, "END_TO_END_FILE", end_to_end), \
-                 patch.object(gen, "BODY_CLOSURE_FILE", body_closure), \
-                 patch.object(
-                     gen,
-                     "_parse_bridge_lemmas",
-                     return_value=(["add"], []),
-                 ):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "universal-safe-body-closure")
-        self.assertEqual(
-            phase4["layers2_3_ir_matches_yul_evmYulLeanBackend"],
-            "proven (conditional on bridged IR bodies)",
-        )
-
-    def test_body_closure_without_end_to_end_backend_wrapper_stays_universal(self) -> None:
-        """EndToEnd no longer needs the backend-wrapper theorem for the
-        universal safe-body closure milestone."""
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            tmp_path = Path(tmp)
-            retarget = tmp_path / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(self._retarget_all_proven(), encoding="utf-8")
-
-            end_to_end = tmp_path / "EndToEnd.lean"
-            end_to_end.write_text(
-                textwrap.dedent("""\
-                    theorem layers2_3_ir_matches_native_evmYulLean : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            body_closure = tmp_path / "EvmYulLeanBodyClosure.lean"
-            body_closure.write_text(
-                textwrap.dedent("""\
-                    theorem compileStmtList_always_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            with patch.object(gen, "RETARGET_FILE", retarget), \
-                 patch.object(gen, "END_TO_END_FILE", end_to_end), \
-                 patch.object(gen, "BODY_CLOSURE_FILE", body_closure), \
-                 patch.object(
-                     gen,
-                     "_parse_bridge_lemmas",
-                     return_value=(["add"], []),
-                 ):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertEqual(phase4["status"], "universal-safe-body-closure")
-        self.assertEqual(
-            phase4["layers2_3_ir_matches_yul_evmYulLeanBackend"],
-            "removed from EndToEnd surface",
-        )
-
-    def test_universal_body_closure_with_admitted_deps_does_not_flip(self) -> None:
-        """If compileStmtList_always_bridged is proven but sar remains
-        admitted, phase4 status must not advance to full_semantic_integration."""
-        with tempfile.TemporaryDirectory(dir=gen.ROOT) as tmp:
-            tmp_path = Path(tmp)
-            retarget = tmp_path / ("EvmYulLean" + "Retarget.lean")
-            retarget.write_text(self._retarget_all_proven(), encoding="utf-8")
-
-            end_to_end = tmp_path / "EndToEnd.lean"
-            end_to_end.write_text(
-                textwrap.dedent("""\
-                    theorem layers2_3_ir_matches_yul_evmYulLeanBackend : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            body_closure = tmp_path / "EvmYulLeanBodyClosure.lean"
-            body_closure.write_text(
-                textwrap.dedent("""\
-                    theorem compileStmtList_always_bridged : True := by
-                      trivial
-                """),
-                encoding="utf-8",
-            )
-
-            with patch.object(gen, "RETARGET_FILE", retarget), \
-                 patch.object(gen, "END_TO_END_FILE", end_to_end), \
-                 patch.object(gen, "BODY_CLOSURE_FILE", body_closure), \
-                 patch.object(
-                     gen,
-                     "_parse_bridge_lemmas",
-                     return_value=(["add", "smod", "sar"], ["smod", "sar"]),
-                 ):
-                report = gen.build_report()
-        phase4 = report["phase4_retarget"]
-        self.assertNotEqual(phase4["status"], "full_semantic_integration")
-        self.assertEqual(phase4["admitted_bridge_dependencies"], ["sar", "smod"])
-        self.assertIn("proven", phase4["compileStmtList_always_bridged"])
-
-    def test_schema_version_bumped_for_universal_body_closure_field(self) -> None:
-        """Adding `compileStmtList_always_bridged` to the phase4 payload bumps
-        the top-level schema_version so downstream consumers can detect the
-        extension."""
-        report = gen.build_report()
-        self.assertEqual(report["schema_version"], 7)
+        self.assertEqual(report["schema_version"], 8)
 
 
 class ParseContextBridgeLemmasTests(unittest.TestCase):
