+++ scripts/verify_sync_spec.json
@@ -697,7 +697,7 @@
     "python3 scripts/check_yul.py --builtin-boundary-only",
     "python3 scripts/check_rewrite_proof_metadata.py",
     "python3 scripts/generate_evmyullean_capability_report.py --check",
-    "python3 scripts/generate_evmyullean_adapter_report.py --check",
+    "python3 scripts/generate_evmyullean_native_lowering_report.py --check",
     "python3 scripts/generate_print_axioms.py --check",
     "python3 scripts/check_proof_length.py",
     "python3 scripts/check_issue_1060_integrity.py",
